requirements.txt
torch~=1.13.1
numpy~=1.23.5
pandas~=1.5.3
mlh~=0.1
scikit-learn~=1.1.3
runx~=0.0.11
tqdm~=4.64.1
scipy~=1.10.0
art~=6.0
torchvision~=0.14.1
matplotlib~=3.7.4
seaborn~=0.12.2
PyYAML~=6.0
opacus~=1.4.0
GPUtil~=1.4.0
Pillow~=9.4.0
setuptools~=65.6.3